added proposal slots to validator duties
