+++ data/provider/src/main/java/tech/pegasys/artemis/api/ChainDataProvider.java
@@ -16,19 +16,13 @@ package tech.pegasys.artemis.api;
 import static tech.pegasys.artemis.datastructures.util.BeaconStateUtil.get_committee_count_at_slot;
 import static tech.pegasys.artemis.util.async.SafeFuture.completedFuture;
 
-import com.google.common.annotations.VisibleForTesting;
 import com.google.common.primitives.UnsignedLong;
-import java.util.ArrayList;
-import java.util.HashMap;
 import java.util.List;
-import java.util.Map;
 import java.util.Optional;
 import java.util.stream.Collectors;
 import org.apache.tuweni.bytes.Bytes32;
-import org.apache.tuweni.bytes.Bytes48;
 import tech.pegasys.artemis.api.schema.Attestation;
 import tech.pegasys.artemis.api.schema.AttestationData;
-import tech.pegasys.artemis.api.schema.BLSPubKey;
 import tech.pegasys.artemis.api.schema.BLSSignature;
 import tech.pegasys.artemis.api.schema.BeaconChainHead;
 import tech.pegasys.artemis.api.schema.BeaconHead;
@@ -37,27 +31,18 @@ import tech.pegasys.artemis.api.schema.BeaconValidators;
 import tech.pegasys.artemis.api.schema.Committee;
 import tech.pegasys.artemis.api.schema.Fork;
 import tech.pegasys.artemis.api.schema.SignedBeaconBlock;
-import tech.pegasys.artemis.api.schema.ValidatorDuties;
-import tech.pegasys.artemis.api.schema.ValidatorDutiesRequest;
 import tech.pegasys.artemis.api.schema.ValidatorsRequest;
-import tech.pegasys.artemis.datastructures.state.CommitteeAssignment;
 import tech.pegasys.artemis.datastructures.util.AttestationUtil;
 import tech.pegasys.artemis.datastructures.util.BeaconStateUtil;
-import tech.pegasys.artemis.statetransition.BlockProposalUtil;
-import tech.pegasys.artemis.statetransition.StateTransition;
 import tech.pegasys.artemis.storage.ChainStorageClient;
 import tech.pegasys.artemis.storage.CombinedChainDataClient;
 import tech.pegasys.artemis.util.SSZTypes.Bitlist;
 import tech.pegasys.artemis.util.async.SafeFuture;
-import tech.pegasys.artemis.util.bls.BLSPublicKey;
-import tech.pegasys.artemis.util.config.Constants;
 
 public class ChainDataProvider {
   private final CombinedChainDataClient combinedChainDataClient;
 
   private final ChainStorageClient chainStorageClient;
-  private final StateTransition stateTransition = new StateTransition();
-  private final BlockProposalUtil blockCreator = new BlockProposalUtil(stateTransition);
 
   public ChainDataProvider(
       final ChainStorageClient chainStorageClient,
@@ -211,125 +196,32 @@ public class ChainDataProvider {
     return Optional.of(attestation);
   }
 
+  public SafeFuture<Optional<BeaconValidators>> getValidatorsByValidatorsRequest(
+      final ValidatorsRequest request) {
+    UnsignedLong slot =
+        request.epoch == null
+            ? combinedChainDataClient.getBestSlot()
+            : BeaconStateUtil.compute_start_slot_at_epoch(request.epoch);
+
+    return combinedChainDataClient
+        .getStateAtSlot(slot)
+        .thenApply(
+            optionalBeaconState -> {
+              if (optionalBeaconState.isEmpty()) {
+                return Optional.empty();
+              }
+              return Optional.of(
+                  new BeaconValidators(
+                      new BeaconState(optionalBeaconState.get()), request.pubkeys));
+            });
+  }
+
   public boolean isFinalized(SignedBeaconBlock signedBeaconBlock) {
     return combinedChainDataClient.isFinalized(signedBeaconBlock.message.slot);
   }
 
   public boolean isFinalized(UnsignedLong slot) {
     return combinedChainDataClient.isFinalized(slot);
-  }
-
-  public SafeFuture<Optional<BeaconValidators>> getValidatorsByValidatorsRequest(
-      final ValidatorsRequest request) {
-    UnsignedLong slot =
-        request.epoch == null
-            ? chainStorageClient.getBestSlot()
-            : BeaconStateUtil.compute_start_slot_at_epoch(request.epoch);
-
-    return getStateAtSlot(slot)
-        .thenApply(
-            optionalBeaconState -> {
-              if (optionalBeaconState.isEmpty()) {
-                return Optional.empty();
-              }
-              return Optional.of(new BeaconValidators(optionalBeaconState.get(), request.pubkeys));
-            });
-  }
-
-  public SafeFuture<List<ValidatorDuties>> getValidatorDutiesByRequest(
-      final ValidatorDutiesRequest validatorDutiesRequest) {
-
-    if (validatorDutiesRequest == null || !isStoreAvailable()) {
-      return completedFuture(List.of());
-    }
-    final Optional<Bytes32> optionalBlockRoot = getBestBlockRoot();
-    if (optionalBlockRoot.isEmpty()) {
-      return completedFuture(List.of());
-    }
-
-    UnsignedLong epoch = validatorDutiesRequest.epoch;
-    UnsignedLong slot = BeaconStateUtil.compute_start_slot_at_epoch(epoch);
-    final Bytes32 headBlockRoot = optionalBlockRoot.get();
-    return combinedChainDataClient
-        .getStateAtSlot(slot, headBlockRoot)
-        .thenApply(
-            state -> getValidatorDutiesFromState(state.get(), validatorDutiesRequest.pubkeys))
-        .exceptionally(err -> List.of());
-  }
-
-  @VisibleForTesting
-  protected List<ValidatorDuties> getValidatorDutiesFromState(
-      final tech.pegasys.artemis.datastructures.state.BeaconState state,
-      final List<BLSPubKey> pubKeys) {
-    try {
-    final List<ValidatorDuties> dutiesList = new ArrayList<>();
-
-    final List<CommitteeAssignment> committees =
-        combinedChainDataClient.getCommitteesFromState(state, state.getSlot());
-    final UnsignedLong firstSlot = state.getSlot();
-    Map<BLSPubKey, List<UnsignedLong>> proposers = new HashMap<>();
-
-    for (int i = 0; i < Constants.SLOTS_PER_EPOCH; i++) {
-      final UnsignedLong thisSlot = firstSlot.plus(UnsignedLong.valueOf(i));
-      BLSPublicKey publicKey = blockCreator.getProposerForSlot(state, thisSlot);
-      BLSPubKey pubkey = new BLSPubKey(publicKey.toBytes());
-      if (proposers.containsKey(pubkey)) {
-        List<UnsignedLong> proposalSlots = proposers.get(pubkey);
-        proposalSlots.add(thisSlot);
-      } else {
-        proposers.put(pubkey, List.of(thisSlot).stream().collect(Collectors.toList()));
-      }
-    }
-
-    for (final BLSPubKey pubKey : pubKeys) {
-      final Integer validatorIndex = getValidatorIndex(state.getValidators().asList(), pubKey);
-      if (validatorIndex == null) {
-        dutiesList.add(new ValidatorDuties(pubKey, null, null, List.of()));
-      } else {
-        List<UnsignedLong> proposedSlots = proposers.getOrDefault(pubKey, List.of());
-        dutiesList.add(
-            new ValidatorDuties(
-                pubKey,
-                validatorIndex,
-                getCommitteeIndex(committees, validatorIndex),
-                proposedSlots));
-      }
-    }
-    return dutiesList;
-    } catch (Exception e) {
-      System.out.println("exception");
-      e.printStackTrace();
-      return List.of();
-    }
-  }
-
-  @VisibleForTesting
-  protected static Integer getValidatorIndex(
-      final List<tech.pegasys.artemis.datastructures.state.Validator> validators,
-      final BLSPubKey publicKey) {
-    Optional<tech.pegasys.artemis.datastructures.state.Validator> optionalValidator =
-        validators.stream()
-            .filter(
-                v -> Bytes48.fromHexString(publicKey.toHexString()).equals(v.getPubkey().toBytes()))
-            .findFirst();
-    if (optionalValidator.isPresent()) {
-      return validators.indexOf(optionalValidator.get());
-    } else {
-      return null;
-    }
-  }
-
-  @VisibleForTesting
-  protected Integer getCommitteeIndex(List<CommitteeAssignment> committees, int validatorIndex) {
-    Optional<CommitteeAssignment> matchingCommittee =
-        committees.stream()
-            .filter(committee -> committee.getCommittee().contains(validatorIndex))
-            .findFirst();
-    if (matchingCommittee.isPresent()) {
-      return committees.indexOf(matchingCommittee.get());
-    } else {
-      return null;
-    }
   }
 
   public Optional<BeaconChainHead> getHeadState() {
